update tarea 2

+++ Tarea-2-Grupo33/src/controllers/DefensaReinoController.js
@@ -1,16 +1,22 @@
+import { parse } from 'dotenv';
 import prisma from '../prismaClient.js';
 
 //relacion defensa-reino
 //PK, FK1: id_reino Integer
 //PK, FK2: id_defensa Integer
 
-
+//=============================================//
+//Create Relacion Defensa-Reino                //
+//=============================================//
 const createDefensaReino = async (req, res, next) => {
+    //Parametros
     const { id_reino, id_defensa } = req.body;
+    //Validacion de parametros
+    if (!id_reino || !id_defensa || isNaN(parseInt(id_reino)) || isNaN(parseInt(id_defensa))) {
+        next({ message: 'Bad request', status: 400 });
+    }
+    //Creacion
     try {
-        if (!id_reino || !id_defensa) {
-            throw new Error('Bad request');
-        }
         const newRelDefRein = await prisma.reino_defensas.create({
             data: {
                 id_reino,
@@ -28,19 +34,38 @@ const createDefensaReino = async (req, res, next) => {
         next(error);
     }
 }
-
+//=============================================//
+//Get Relacion Defensa-Reino                   //
+//=============================================//
 const getDefensaReino = async (req, res, next) => {
+    //Busqueda
     try {
         const rel_def_reins = await prisma.reino_defensas.findMany();
+        // Si retorna [] es un not found
+        if (rel_def_reins.length === 0) {
+            throw new Error('Not Found');
+        }
         res.status(200).json(rel_def_reins); //OK
     } catch (error) {
+        if (error.message === 'Not Found') {
+            error.status = 404;
+        } else {
         error.status = 500; // Internal Server Error
+        }
         next(error);
     }
 }
-
+//=============================================//
+//Get Relacion Defensa-Reino by id             //
+//=============================================//
 const getDefensaReinoById = async (req, res, next) => {
+    //Parametros
     const { id } = req.params;
+    //Validacion de parametros
+    if (!id || isNaN(parseInt(id))) {
+        next({ message: 'Bad request', status: 400 });
+    }
+    //Busqueda
     try {
         const rel_def_rein = await prisma.reino_defensas.findUnique({
             where: {
@@ -63,10 +88,18 @@ const getDefensaReinoById = async (req, res, next) => {
 
     }
 }
-
+//=================================================================//
+//Update Relacion Defensa-Reino                                    //
+//=================================================================//
 const updateDefensaReino = async (req, res, next) => {
+    //Parametros
     const { id } = req.params;
     const { id_reino, id_defensa } = req.body;
+    //Validacion de parametros
+    if (!id || !id_reino || !id_defensa || isNaN(parseInt(id)) || isNaN(parseInt(id_reino)) || isNaN(parseInt(id_defensa))) {
+        next({ message: 'Bad request', status: 400 });
+    }
+    //Actualizacion
     try {
         const updatedRelDefRein = await prisma.reino_defensas.update({
             where: {
@@ -86,7 +119,13 @@ const updateDefensaReino = async (req, res, next) => {
 }
 
 const deleteDefensaReino = async (req, res, next) => {
+    //Parametros
     const { id } = req.params;
+    //Validacion de parametros
+    if (!id || isNaN(parseInt(id))) {
+        next({ message: 'Bad request', status: 400 });
+    }
+    //Eliminacion
     try {
         const deletedRelDefRein = await prisma.reino_defensas.delete({
             where: {
@@ -103,11 +142,10 @@ const deleteDefensaReino = async (req, res, next) => {
             error.status = 500; // Internal Server Error
         }
         next(error);
-
     }
 }
 
-
+//=============================================//
 const DefensaReinoController = {
     createDefensaReino,
     getDefensaReino,
@@ -115,5 +153,6 @@ const DefensaReinoController = {
     updateDefensaReino,
     deleteDefensaReino
 }
+//=============================================//
 
 export default DefensaReinoController;
